Move local/global variable functions to variables runtime module
